pyproject.toml
[project]
name = "ilivery"
dynamic = ["version"]
description = ""
authors = [
    {name = "Logan Grado", email = "grado.logan@gmail.com"},
]
dependencies = [
    "matplotlib<4.0.0,>=3.5.3",
    "numpy<2.0.0,>=1.23.2",
    "PyQt5<6.0.0,>=5.15.7",
    "psd-tools<2.0.0,>=1.9.21",
    "Pillow<10.0.0,>=9.2.0",
    "click>=8.1.7",
    "opensimplex<1.0.0,>=0.4.4",
    "pandas<3.0.0,>=2.0.2",
    "pydantic>=2.8.2",
    "jsonnet>=0.20.0",
    "loggerado>=0.2.0",
]
requires-python = ">=3.10,<4.0"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ipython<9.0.0,>=8.4.0",
    "ipdb<1.0.0,>=0.13.9",
    "black<23.0.0,>=22.6.0",
    "flake8<6.0.0,>=5.0.4",
    "pytest<8.0.0,>=7.1.3",
]

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "scm"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.black]
line-length=120

[project.scripts]
ilivery = "ilivery.cli:main"
